rework replica code
