+++ tembo-operator/Dockerfile
@@ -1,12 +1,12 @@
 # This does not build on ARM laptops, even with the --platform flag.
 
-FROM --platform=linux/amd64 quay.io/tembo/muslrust:1.71.0-stable AS builder
+FROM --platform=linux/amd64 quay.io/tembo/muslrust:1.77.0-stable AS builder
 
 WORKDIR /build
 
 COPY . .
 
-RUN cargo build --release --target=x86_64-unknown-linux-musl
+RUN cargo build --release --target=x86_64-unknown-linux-musl --features=telemetry
 
 FROM --platform=linux/amd64 quay.io/tembo/alpine:3.18.2
