+++ tembo-operator/src/cloudnativepg/cnpg.rs
@@ -8,10 +8,7 @@ use crate::{
         postgres_parameters::MergeError,
     },
     cloudnativepg::{
-        backups::{
-            Backup, BackupCluster, BackupMethod, BackupOnlineConfiguration, BackupSpec,
-            BackupTarget,
-        },
+        backups::Backup,
         clusters::{
             Cluster, ClusterAffinity, ClusterBackup, ClusterBackupBarmanObjectStore,
             ClusterBackupBarmanObjectStoreData, ClusterBackupBarmanObjectStoreDataCompression,
@@ -46,6 +43,7 @@ use crate::{
             ClusterServiceAccountTemplate, ClusterServiceAccountTemplateMetadata, ClusterSpec,
             ClusterStorage, ClusterSuperuserSecret,
         },
+        cnpg_backups::create_backup_if_needed,
         poolers::{
             Pooler, PoolerCluster, PoolerPgbouncer, PoolerSpec, PoolerTemplate, PoolerTemplateSpec,
             PoolerTemplateSpecContainers, PoolerType,
@@ -65,14 +63,11 @@ use crate::{
     trunk::extensions_that_require_load,
     Context, RESTARTED_AT,
 };
-use chrono::{DateTime, NaiveDateTime, Offset, Utc};
+use chrono::{DateTime, NaiveDateTime, Offset};
 use k8s_openapi::api::core::v1::Service;
 use k8s_openapi::apimachinery::pkg::util::intstr::IntOrString;
-use k8s_openapi::{
-    api::core::v1::Pod, apimachinery::pkg::apis::meta::v1::ObjectMeta,
-    apimachinery::pkg::apis::meta::v1::Time,
-};
-use kube::api::{ObjectList, PostParams};
+use k8s_openapi::{api::core::v1::Pod, apimachinery::pkg::apis::meta::v1::ObjectMeta};
+use kube::api::PostParams;
 use kube::{
     api::{DeleteParams, ListParams, Patch, PatchParams},
     runtime::{controller::Action, wait::Condition},
@@ -1190,188 +1185,6 @@ pub async fn reconcile_cnpg(cdb: &CoreDB, ctx: Arc<Context>) -> Result<(), Actio
     Ok(())
 }
 
-async fn create_backup_if_needed(
-    cdb: &CoreDB,
-    ctx: &Arc<Context>,
-    cluster: &Cluster,
-) -> Result<(), Action> {
-    let name = cdb.name_any();
-    let namespace = cluster.metadata.namespace.as_ref().ok_or_else(|| {
-        error!(
-            "Cluster namespace is empty for instance: {}.",
-            name.as_str()
-        );
-        Action::requeue(tokio::time::Duration::from_secs(300))
-    })?;
-
-    // We need to check if the cdb.replicas is changing from 1 -> 2.  If it is, we need to take a
-    // snapshot if snapshots are enabled.  We will need to wait on the snapshot to complete before
-    // allowing the new Cluster configuration to be applied.
-    if !replicas_increasing(cdb, cluster) {
-        return Ok(());
-    }
-
-    // check if snapshots are enabled, if not return OK
-    if let Some(volume_snapshot) = &cdb.spec.backup.volume_snapshot {
-        if !volume_snapshot.enabled {
-            return Ok(());
-        }
-    } else {
-        return Ok(());
-    }
-
-    // Check when last backup was taken from Cluster.status.lastSuccessfulBackup.
-    // If longer than 60 minutes and snapshots_enabled is true then take a snapshot
-    let last_backup = cluster
-        .status
-        .as_ref()
-        .and_then(|s| s.last_successful_backup.as_ref())
-        .and_then(|l| l.parse::<DateTime<Utc>>().ok());
-
-    let now = Utc::now();
-    let duration = now.signed_duration_since(last_backup.unwrap_or(now));
-    if duration.num_minutes() <= 60 {
-        return Ok(());
-    }
-
-    // At this point, we know replicas are increasing, snapshots are enabled, and the last backup was taken over 60 minutes ago
-
-    let backup_api: Api<Backup> = Api::namespaced(ctx.client.clone(), namespace);
-
-    // List all backups
-    // TODO: filter by cluster name in case we have multiple in the same namespace in the future for some reason
-    let lp = ListParams::default();
-    let backups = backup_api.list(&lp).await.map_err(|e| {
-        error!("Error listing backups: {}", e);
-        Action::requeue(Duration::from_secs(300))
-    })?;
-
-    let currently_running_volume_snaps = has_currently_running_volume_snaps(&backups, Time(now));
-
-    if currently_running_volume_snaps {
-        warn!("Active backups detected, requeuing in 30 seconds");
-        return Err(Action::requeue(Duration::from_secs(30)));
-    }
-
-    create_replica_snapshot(cdb, ctx.clone()).await?;
-
-    info!(
-        "Created a new backup for {}, requeuing in 30 seconds",
-        name.as_str()
-    );
-    // Remove the `return` keyword here
-    Err(Action::requeue(Duration::from_secs(30)))
-}
-
-fn has_currently_running_volume_snaps(backups: &ObjectList<Backup>, now: Time) -> bool {
-    backups.items.iter().any(|b| {
-        (b.status
-            .as_ref()
-            .and_then(|s| s.phase.as_deref())
-            .map_or(true, |phase| {
-                phase.is_empty() || matches!(phase, "running" | "pending" | "finishing")
-            }))
-            && b.metadata
-                .creation_timestamp
-                .as_ref()
-                .map_or(false, |creation_time| {
-                    let duration = now.0.signed_duration_since(creation_time.0);
-                    duration.num_minutes() <= 60
-                })
-            && matches!(b.spec.method.as_ref(), Some(BackupMethod::VolumeSnapshot))
-    })
-}
-
-// create_replica_snapshot creates a snapshot (backup) of the current primary instance
-// so we can create a new replica from it.
-#[instrument(skip(cdb, ctx), fields(trace_id, instance_name = %cdb.name_any()))]
-async fn create_replica_snapshot(cdb: &CoreDB, ctx: Arc<Context>) -> Result<(), Action> {
-    let name = cdb.name_any();
-    let namespace = cdb.metadata.namespace.as_ref().ok_or_else(|| {
-        error!("CoreDB namespace is empty for instance: {}.", name);
-        Action::requeue(tokio::time::Duration::from_secs(300))
-    })?;
-
-    // Setup the API to Backup
-    let backup_api: Api<Backup> = Api::namespaced(ctx.client.clone(), namespace);
-
-    // Generate snapshot name.  Should be name + replica + date (replica-20240327045600) but also less than 54 characters
-    let snapshot_name = generate_snapshot_name(&name);
-
-    // Gererate the Backup object
-    let backup = Backup {
-        metadata: ObjectMeta {
-            name: Some(snapshot_name),
-            namespace: Some(namespace.to_string()),
-            ..ObjectMeta::default()
-        },
-        spec: BackupSpec {
-            cluster: BackupCluster { name: name.clone() },
-            method: Some(BackupMethod::VolumeSnapshot),
-            online: Some(true),
-            online_configuration: Some(BackupOnlineConfiguration {
-                immediate_checkpoint: Some(true),
-                ..BackupOnlineConfiguration::default()
-            }),
-            target: Some(BackupTarget::Primary),
-        },
-        status: None,
-    };
-
-    // Apply the new backup object
-    let ps = PatchParams::apply("cntrlr").force();
-
-    let _ = backup_api
-        .patch(&name, &ps, &Patch::Apply(&backup))
-        .await
-        .map_err(|e| {
-            error!("Error patching backup: {}", e);
-            Action::requeue(Duration::from_secs(300))
-        })?;
-
-    return Ok(());
-}
-
-// generate_snapshot_name generates a snapshot name based on the instance name and the current timestamp
-fn generate_snapshot_name(name: &str) -> String {
-    let now = Utc::now();
-    let timestamp = now.format("%Y%m%d%H%M%S").to_string();
-
-    let max_name_len = 54 - 9 - timestamp.len();
-    let truncated_name = if name.len() > max_name_len {
-        &name[..max_name_len]
-    } else {
-        name
-    };
-
-    format!("{}-replica-{}", truncated_name, timestamp)
-}
-
-// Checks to see if the instance count is changing from 1 -> 2
-fn replicas_increasing(cdb: &CoreDB, cluster: &Cluster) -> bool {
-    // Desired replicas from the CoreDB object
-    let cdb_replicas: i64 = cdb.spec.replicas.into();
-
-    // Current instances from Cluster status
-    let current_instances = cluster.status.as_ref().and_then(|s| s.instances);
-
-    // Check if transitioning from 1 to 2
-    if matches!((cdb_replicas, current_instances), (2, Some(1))) {
-        info!(
-            "Instance count is changing from 1 to 2 for instance {}. Desired cdb_replicas: {}, Current instances: {}",
-            cdb.name_any(), cdb_replicas, current_instances.unwrap()
-        );
-        return true;
-    } else {
-        debug!(
-            "No transition from 1 to 2 detected for instance {}. Desired cdb_replicas: {}, Current instances: {}",
-            cdb.name_any(), cdb_replicas, current_instances.unwrap()
-        );
-    }
-
-    false
-}
-
 pub async fn reconcile_metrics_ingress_route(
     cdb: &CoreDB,
     ctx: Arc<Context>,
@@ -2403,16 +2216,7 @@ async fn does_cluster_exist(cdb: &CoreDB, ctx: Arc<Context>) -> Result<bool, Act
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        apis::coredb_types::{CoreDB, CoreDBSpec},
-        cloudnativepg::{
-            backups::{Backup, BackupSpec, BackupStatus},
-            clusters::{Cluster, ClusterSpec, ClusterStatus},
-        },
-    };
-    use chrono::{TimeDelta, Utc};
-    use k8s_openapi::apimachinery::pkg::apis::meta::v1::ObjectMeta;
-    use kube::api::ListMeta;
+    use crate::{apis::coredb_types::CoreDB, cloudnativepg::clusters::Cluster};
     use serde_json::json;
     use std::collections::BTreeMap;
 
@@ -3230,216 +3034,5 @@ mod tests {
             generate_scheduled_backup_snapshot_name(short_name),
             "stormy-capybara-snap"
         );
-    }
-
-    #[test]
-    fn test_check_cluster_instance_count() {
-        // Case 1: Scaling from 1 to 2 instances
-        let cdb = CoreDB {
-            metadata: ObjectMeta {
-                name: Some("test-cluster".to_string()),
-                namespace: Some("test".to_string()),
-                ..ObjectMeta::default()
-            },
-            spec: CoreDBSpec {
-                replicas: 2,
-                ..CoreDBSpec::default()
-            },
-            status: None,
-        };
-        let cluster = Cluster {
-            metadata: ObjectMeta {
-                name: Some("test-cluster".to_string()),
-                namespace: Some("test".to_string()),
-                ..ObjectMeta::default()
-            },
-            spec: ClusterSpec {
-                instances: 2,
-                ..ClusterSpec::default()
-            },
-            status: Some(ClusterStatus {
-                instances: Some(1),
-                ..ClusterStatus::default()
-            }),
-        };
-        assert!(replicas_increasing(&cdb, &cluster));
-
-        // Case 2: No scaling, current instances already 2
-        let cdb = CoreDB {
-            metadata: ObjectMeta {
-                name: Some("test-cluster".to_string()),
-                namespace: Some("test".to_string()),
-                ..ObjectMeta::default()
-            },
-            spec: CoreDBSpec {
-                replicas: 2,
-                ..CoreDBSpec::default()
-            },
-            status: None,
-        };
-        let cluster = Cluster {
-            metadata: ObjectMeta {
-                name: Some("test-cluster".to_string()),
-                namespace: Some("test".to_string()),
-                ..ObjectMeta::default()
-            },
-            spec: ClusterSpec {
-                instances: 2,
-                ..ClusterSpec::default()
-            },
-            status: Some(ClusterStatus {
-                instances: Some(2),
-                ..ClusterStatus::default()
-            }),
-        };
-        assert!(!replicas_increasing(&cdb, &cluster));
-
-        // Case 3: No scaling, replicas and instances not 2
-        let cdb = CoreDB {
-            metadata: ObjectMeta {
-                name: Some("test-cluster".to_string()),
-                namespace: Some("test".to_string()),
-                ..ObjectMeta::default()
-            },
-            spec: CoreDBSpec {
-                replicas: 1,
-                ..CoreDBSpec::default()
-            },
-            status: None,
-        };
-        let cluster = Cluster {
-            metadata: ObjectMeta {
-                name: Some("test-cluster".to_string()),
-                namespace: Some("test".to_string()),
-                ..ObjectMeta::default()
-            },
-            spec: ClusterSpec {
-                instances: 1,
-                ..ClusterSpec::default()
-            },
-            status: Some(ClusterStatus {
-                instances: Some(1),
-                ..ClusterStatus::default()
-            }),
-        };
-        assert!(!replicas_increasing(&cdb, &cluster));
-    }
-
-    #[test]
-    fn test_has_currently_running_volume_snaps() {
-        let now = Utc::now();
-
-        let backup_running = Backup {
-            metadata: ObjectMeta {
-                creation_timestamp: Some(Time(now - TimeDelta::try_minutes(60).unwrap())),
-                ..ObjectMeta::default()
-            },
-            spec: BackupSpec {
-                method: Some(BackupMethod::VolumeSnapshot),
-                ..BackupSpec::default()
-            },
-            status: Some(BackupStatus {
-                phase: Some("running".to_string()),
-                ..BackupStatus::default()
-            }),
-        };
-
-        let backup_pending = Backup {
-            metadata: ObjectMeta {
-                creation_timestamp: Some(Time(now - TimeDelta::try_minutes(60).unwrap())),
-                ..ObjectMeta::default()
-            },
-            spec: BackupSpec {
-                method: Some(BackupMethod::VolumeSnapshot),
-                ..BackupSpec::default()
-            },
-            status: Some(BackupStatus {
-                phase: Some("pending".to_string()),
-                ..BackupStatus::default()
-            }),
-        };
-
-        let backup_finishing = Backup {
-            metadata: ObjectMeta {
-                creation_timestamp: Some(Time(now - TimeDelta::try_minutes(60).unwrap())),
-                ..ObjectMeta::default()
-            },
-            spec: BackupSpec {
-                method: Some(BackupMethod::VolumeSnapshot),
-                ..BackupSpec::default()
-            },
-            status: Some(BackupStatus {
-                phase: Some("finishing".to_string()),
-                ..BackupStatus::default()
-            }),
-        };
-
-        let backup_completed = Backup {
-            metadata: ObjectMeta {
-                creation_timestamp: Some(Time(now - TimeDelta::try_minutes(90).unwrap())),
-                ..ObjectMeta::default()
-            },
-            spec: BackupSpec {
-                method: Some(BackupMethod::VolumeSnapshot),
-                ..BackupSpec::default()
-            },
-            status: Some(BackupStatus {
-                phase: Some("completed".to_string()),
-                ..BackupStatus::default()
-            }),
-        };
-
-        // Check if there is a currently running volume snapshot
-        let backups_list = ObjectList {
-            items: vec![backup_running.clone(), backup_completed.clone()],
-            metadata: ListMeta::default(),
-        };
-
-        assert!(has_currently_running_volume_snaps(&backups_list, Time(now)));
-
-        // Check if there is a currently pending volume snapshot
-        let backups_list = ObjectList {
-            items: vec![backup_pending.clone(), backup_completed.clone()],
-            metadata: ListMeta::default(),
-        };
-
-        assert!(has_currently_running_volume_snaps(&backups_list, Time(now)));
-
-        // Check if there is a currently finishing volume snapshot
-        let backups_list = ObjectList {
-            items: vec![backup_finishing.clone(), backup_completed.clone()],
-            metadata: ListMeta::default(),
-        };
-
-        assert!(has_currently_running_volume_snaps(&backups_list, Time(now)));
-
-        let backups_list = ObjectList {
-            items: vec![backup_completed.clone()],
-            metadata: ListMeta::default(),
-        };
-
-        assert!(!has_currently_running_volume_snaps(
-            &backups_list,
-            Time(now)
-        ));
-    }
-
-    #[test]
-    fn test_generate_snapshot_name() {
-        // Test case 1: Name fits within 54 characters
-        let name1 = "my-snapshot";
-        let snapshot_name1 = generate_snapshot_name(name1);
-        assert!(snapshot_name1.starts_with(name1));
-        assert!(snapshot_name1.ends_with(&Utc::now().format("%Y%m%d%H%M%S").to_string()));
-        assert!(snapshot_name1.len() <= 54);
-
-        // Test case 2: Name exceeds 54 characters
-        let name2 = "a-very-long-snapshot-name-that-exceeds-54-characters";
-        let snapshot_name2 = generate_snapshot_name(name2);
-        let timestamp_len = Utc::now().format("%Y%m%d%H%M%S").to_string().len();
-        let max_name_len = 54 - 9 - timestamp_len;
-        assert!(snapshot_name2.starts_with(&name2[..max_name_len]));
-        assert!(snapshot_name2.ends_with(&Utc::now().format("%Y%m%d%H%M%S").to_string()));
-        assert!(snapshot_name2.len() <= 54);
     }
 }
